build: add experimental support for RISCV
This adds an experimental path to support RISCV targets with the vector
extensions.  This path assumes the use of clang/LLVM as there are clang
specific options in play here to enable the vector extensions.  Support
for GCC should be an easy extension if desired, but can be handled as a
separate change.

PiperOrigin-RevId: 392973972